Api noc dev to prod

+++ nocWebApi/Models/MedicinalIngredient.cs
@@ -10,7 +10,7 @@ namespace nocWebApi.Models
         public int noc_number { get; set; }
         public int noc_pi_din_product_id { get; set; }
         public string noc_pi_medic_ingr_name { get; set; }
-        public int noc_pi_strength { get; set; }
+        public double noc_pi_strength { get; set; }
         public string noc_pi_unit { get; set; }
         public string noc_pi_basic_unit { get; set; }
     }
